Rm unecessary cast

+++ readers/influxdb/messages_test.go
@@ -70,7 +70,7 @@ func TestReadAll(t *testing.T) {
 	for i := 0; i < msgsNum; i++ {
 		// Mix possible values as well as value sum.
 		msg := m
-		msg.Time = float64(now)/float64(1e9) - float64(i)
+		msg.Time = float64(now)/1e9 - float64(i)
 
 		count := i % valueFields
 		switch count {
